+++ plugins/LiveTemplatesWithSelection/src/main/kotlin/dev/ghostflyby/selectionlivetemplate/KeyDelegate.kt
@@ -1,0 +1,28 @@
+/*
+ * Copyright (c) 2025 ghostflyby <ghostflyby+intellij@outlook.com>
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3.0 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, see
+ * <https://www.gnu.org/licenses/>.
+ */
+
+package dev.ghostflyby.selectionlivetemplate
+import com.intellij.openapi.util.Key
+import com.intellij.openapi.util.UserDataHolder
+import kotlin.reflect.KProperty
+
+internal operator fun <T : Any> Key<T>.getValue(thisRef: UserDataHolder, property: KProperty<*>): T? =
+    thisRef.getUserData(this)
+
+internal operator fun <T : Any> Key<T>.setValue(thisRef: UserDataHolder, property: KProperty<*>, value: T?) =
+    thisRef.putUserData(this, value)