+++ plugins/LiveTemplatesWithSelection/CHANGELOG.md
@@ -1,7 +1,8 @@
 <!-- Keep a Changelog guide -> https://keepachangelog.com -->
 
-# IntelliJ-Plugins Changelog
+# LiveTemplatesWithSelection Changelog
 
 ## [Unreleased]
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+- Implementing the main functionality to expand `$SELECTION$`